+++ tests/wpunit/Traits/__snapshots__/Custom_Table_Query_MethodsTest__should_count_total_items_with_sub_where_clauses__0.snapshot.php
@@ -1,0 +1,1 @@
+SELECT COUNT(*) FROM `wp_query_test` a WHERE (a.slug = 'x' OR a.slug = 'y')