+++ tests/wpunit/Traits/__snapshots__/Custom_Table_Query_MethodsTest__should_handle_deeply_nested_sub_wheres_four_levels__0.snapshot.php
@@ -1,0 +1,1 @@
+SELECT a.* FROM `wp_query_test` a  WHERE (((a.slug = 'a1' OR a.slug = 'a2') AND a.status = 1) OR a.slug = 'b1') ORDER BY a.id ASC LIMIT 0, 10