+++ tests/wpunit/Traits/__snapshots__/Custom_Table_Query_MethodsTest__should_paginate_with_nested_sub_where_clauses__0.snapshot.php
@@ -1,0 +1,1 @@
+SELECT a.* FROM `wp_query_test` a  WHERE (a.slug = 'alpha' OR a.slug = 'beta') AND a.status = 1 ORDER BY a.id ASC LIMIT 0, 10