+++ tests/wpunit/Traits/__snapshots__/Custom_Table_Query_MethodsTest__should_produce_correct_query_for_simple_sub_where__0.snapshot.php
@@ -1,0 +1,1 @@
+SELECT a.* FROM `wp_query_test` a  WHERE a.slug = 'test' ORDER BY a.id ASC LIMIT 0, 20